Move demo file away from components
