+++ src/Demo.tsx
@@ -1,11 +1,11 @@
 import * as React from "react";
-import Panel from "./Panel";
-import Header from "./Header";
-import Form from "./Form";
-import Fieldset from "./Fieldset";
-import Input from "./Input";
-import Tabs from "./Tabs";
-import SubmitButton from "./SubmitButton";
+import Panel from "./components/Panel";
+import Header from "./components/Header";
+import Form from "./components/Form";
+import Fieldset from "./components/Fieldset";
+import Input from "./components/Input";
+import Tabs from "./components/Tabs";
+import SubmitButton from "./components/SubmitButton";
 
 export interface DemoState {
   lights: boolean;
