+++ src/index.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
 import * as ReactDOM from "react-dom";
-import Demo from "./components/Demo";
+import Demo from "./Demo";
 
 ReactDOM.render(<Demo  />, document.getElementById("main"));
